[refactor] Add test_shell.cpp
