--- Testshell/test.cpp
@@ -1,5 +1,0 @@
-#include "main.cpp"
-
-TEST(SampleTest, HandlesTrue) {
-    EXPECT_TRUE(true);
-}
